Dzień 16

+++ ChallengeApp/ChallengeApp/IEmployee.cs
@@ -12,14 +12,12 @@
 
         void AddGrade(double grade);
 
-        void AddGrade(long grade);
+        void AddGrade(int grade);
 
         void AddGrade(char grade);
 
         void AddGrade(string grade);
 
-
         Statistics GetStatistics();
-
     }
 }
